requirements.txt
# VFL-RAG: VFL/SHAP training & prediction, RAG vector store + LLM action plans, scoring notebooks
# Install: pip install -r requirements.txt

# -----------------------------------------------------------------------------
# Core numerics & tables
# -----------------------------------------------------------------------------
numpy>=1.23.0
pandas>=1.5.0
joblib>=1.2.0

# -----------------------------------------------------------------------------
# Deep learning (Train/Predict notebooks, utils.model_utils, RAG summarization GPU)
# -----------------------------------------------------------------------------
torch>=2.0.0

# -----------------------------------------------------------------------------
# ML, explainability, training (Train / Predict notebooks)
# -----------------------------------------------------------------------------
scikit-learn>=1.2.0
shap>=0.42.0

# -----------------------------------------------------------------------------
# Notebooks (Jupyter Notebook UI + kernel for .venv)
# -----------------------------------------------------------------------------
tqdm>=4.65.0
ipywidgets>=8.0.0
ipykernel>=6.29.0
notebook>=7.0.0

# -----------------------------------------------------------------------------
# Environment & LLM APIs (RAG Part 2, optional LLM titles in rag_index_build)
# -----------------------------------------------------------------------------
python-dotenv>=1.0.0
openai>=1.0.0

# -----------------------------------------------------------------------------
# LangChain RAG (Index / Plan notebooks: FAISS, loaders, splitters, embeddings)
# On Python 3.14+, LangChain still touches pydantic.v1 and triggers a UserWarning;
# utils/rag_utils.py, utils/rag_index_build.py, and Part 1 notebook register a
# narrow filter before importing LangChain. For a fully supported stack, use Python 3.12–3.13.
# On Windows, use Python 3.12 for ColBERT: ``voyager`` (RAGatouille) has no cp313 wheels.
# -----------------------------------------------------------------------------
langchain>=0.3.0
langchain-core>=0.1.0
langchain-community>=0.0.1
langchain-classic>=1.0.0,<2.0.0
langchain-openai>=0.2.0
langchain-text-splitters>=0.0.1

# -----------------------------------------------------------------------------
# Vector store & sentence embeddings (FAISS index, MiniLM, etc.)
# -----------------------------------------------------------------------------
faiss-cpu>=1.7.4
# Optional GPU FAISS: pip install faiss-gpu>=1.7.4
sentence-transformers>=2.2.0

# -----------------------------------------------------------------------------
# Retrieval / reranking (Part 2) — six-mode topology used by Plan.ipynb:
#   none               - FAISS dense top-100 + MMR                (sentence-transformers)
#   bm25               - BM25 over full corpus (single-stage)     (rank-bm25)
#   doct5query         - BM25 over T5-expanded corpus             (rank-bm25 + transformers + sentencepiece)
#   crossencoder       - FAISS top-100 -> CrossEncoder            (sentence-transformers)
#   bm25_crossencoder  - BM25 top-100 -> CrossEncoder             (rank-bm25 + sentence-transformers)
#   colbert            - FAISS top-100 -> ColBERT MaxSim          (ragatouille)
# Windows: first ColBERT load JIT-builds a torch extension; install "Desktop development with C++" (MSVC/cl) and keep ``ninja`` on PATH (venv ``Scripts``).
# -----------------------------------------------------------------------------
ragatouille>=0.0.9.post2,<0.0.10
rank-bm25>=0.2.2
# SentencePiece tokenizer is required by the doc2query T5 model
# (``castorini/doc2query-t5-base-msmarco``) used by the ``doct5query`` mode.
sentencepiece>=0.1.99

# -----------------------------------------------------------------------------
# Hugging Face (local summarization titles: distilbart; pulled by sentence-transformers)
# -----------------------------------------------------------------------------
# ColBERT (RAGatouille / colbert-ai) breaks on transformers 5.x (missing all_tied_weights_keys on HF_ColBERT).
transformers>=4.41.0,<5.0.0
huggingface-hub>=0.23.0,<1.0.0
ninja>=1.11.0
safetensors>=0.4.0
protobuf>=3.20.0

# -----------------------------------------------------------------------------
# PDF knowledge base (PyPDFLoader / rag_index_build section extract)
# -----------------------------------------------------------------------------
pypdf>=3.0.0

# -----------------------------------------------------------------------------
# Score.ipynb (BERTScore, BLEU/NLTK, ROUGE, SBERT cosine)
# -----------------------------------------------------------------------------
matplotlib>=3.7.0
seaborn>=0.12.0
bert-score>=0.3.13
nltk>=3.8.0
rouge-score>=0.1.2
